chore: move sharedtypes package to internal
